Default tags only if new present in the image name
