+++ pkg/defaulting/images_test.go
@@ -9,6 +9,8 @@ import (
 	"fmt"
 	"reflect"
 	"testing"
+
+	"github.com/stretchr/testify/assert"
 )
 
 func TestGetLatestAgentImage(t *testing.T) {
@@ -176,5 +178,23 @@ func TestNewImage(t *testing.T) {
 				t.Errorf("NewImage() = %v, want %v", got, tt.want)
 			}
 		})
+	}
+}
+
+func TestIsImageNameContainsTag(t *testing.T) {
+	cases := map[string]bool{
+		"foo:bar":             true,
+		"foo/bar:baz":         true,
+		"foo/bar:baz:tar":     true,
+		"foo/bar:baz-tar":     true,
+		"foo/bar:baz_tar":     true,
+		"foo/bar:baz.tar":     true,
+		"foo/foo/bar:baz:tar": true,
+		"foo":                 false,
+		":foo":                false,
+		"foo:foo/bar":         false,
+	}
+	for tc, expected := range cases {
+		assert.Equal(t, expected, IsImageNameContainsTag(tc))
 	}
 }
